requirements.txt
pylint
pytest
pytest-cov
requests
pandas
beautifulsoup4
numpy
python-craigslist
opentelemetry-api
opentelemetry-instrumentation
Flask==1.1.1
dash==1.12.0
gunicorn
pandas-gbq
google-cloud
google-cloud-bigquery
google-api-python-client
google-cloud-ndb
google-cloud-datastore
google-cloud-storage